add RTDP
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	new file:   RL.code-workspace
	modified:   rl_algorithms/tabular/agents/__init__.py
	new file:   rl_algorithms/tabular/agents/_rtdp.py
	modified:   rl_algorithms/tabular/exploration/_epsilon_greedy.py
	modified:   rl_algorithms/tabular/exploration/_greedy.py
	modified:   rl_algorithms/tabular/exploration/_softmax.py
	modified:   run.py

+++ rl_algorithms/tabular/exploration/_greedy.py
@@ -8,7 +8,7 @@ class Greedy:
 
     def choose_action(self, action_space:gym.spaces.Discrete, state: int, qtable: np.ndarray, mask: np.ndarray | None=None) -> int:
         if mask is None:
-            mask = np.ones_like(qtable)
+            mask = np.ones_like(qtable).astype(np.int8)
 
         action = self.argmax(qtable[state, :], mask[state, :])
